requirements.txt
Django==4.2.9
djangorestframework==3.15.2
psycopg2-binary==2.9.9
python-dotenv==1.0.1
python-telegram-bot==13.15
requests==2.32.4
gunicorn==23.0.0
celery==5.4.0
redis==5.0.8
ipython==8.18.1
geopy==2.4.1
urllib3==1.26.18
pika==1.3.2
Pillow==11.3.0
